+++ lib/src/mock_auth_result.dart
@@ -1,0 +1,8 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:mockito/mockito.dart';
+
+import 'mock_firebase_user.dart';
+
+class MockAuthResult extends Mock implements AuthResult {
+  FirebaseUser user = MockFirebaseUser();
+}
